requirements.txt
selenium==4.9.1
pycryptodome==3.17
pandas==1.5.3
tqdm==4.65.0
requests==2.28.2
cryptography==41.0.2
pillow==9.5.0
